CLI 'show' and 'checksums' command
'show' command:
Starts QIV image viewer for the thumbnails of the given pictures
directory. The function is called show_dir and starts QIV via
subprocess

'checksums' command:
Prints a list of the picture checksums that is compatible with the
output from the sha1sum program.

picture.thumbnail property:
Whenever a sidecar is added to a picture the add_sidecar method checks
if it of type 'Thumbnail'. If it is the thumbnail instance property
is updated with the added path.

minor:
It is now possible to call pic.py directly without specifying python.