Switch to relative paths
Now all paths are always relative paths and the pictures instances only store the filename of the picture files without the path. This makes it possible to easily move a whole tree without having to rewrite the cache and the picture instances.
As a consequence the path to the log dir has also changed. Now the log dir is inside the picture directory (i.e. <picdir>/log/). Most changes were in pic.py which got a helper function delete_tree(root) which removes a directory and all its contennts. This function is used to remove the log directory when cleaning the picture directory (clean_dir).

pic_import.py is finally removed from the repository as it has no valuable content.