Dispatcher thread for Stage workers, CLI progress bar & better logging
Dispatcher (dispatcher.py):
The worker threads in a stage are now controlled by a dispatcher
thread that also belongs to the same Stage. A Dispatcher thread halts
and awakens worker threads when needed: whenever the input queue of the
stage is empty the dispatcher tells the worker threads to wait. As soon
as there are jobs to process the dispatcher activates the worker
threads. The stage no longer directly interacts with its workers and the
worker threads no longer quit on their own if there is no work left
(this was a bug and caused worker in the last stages to terminate before
they even got jobs to work on). The dispatcher also monitors the progress
of the stage and in the future will add or remove worker threads when
needed.

CLI progress bar (pic.py):
The import_dir() function now not only waits for the workers to complete
their jobs but polls the overall progress of the pipeline and visualizes
the progress with a progress bar. The progress bar is implemented with a
new class ProgressBar (in pic.py).

Better logging: (worker.py):
There now exists a new config variable LOGGING=[True/False] that controls
if the worker threads should log their progress to file or not. If logging
is enabled pic.import_dir will ensure that the logging dir exists before
the worker threads are started (before PiC crashed if the directory didn't
exist). The worker threads then constantly log what they are working on.