Reworked Repo class. New RepoFactory methods.
Repo.connect now also loads repo configuration after connecting
to it. Repo.write_default_config has been replaced by
Repo._create_default_config and Repo.write_config.

Two factory methods:
1) RepoFactory.init_dir
initialize a directory intended as a repo directory: create
.pic directory, write default config to file and create an empty
index file.
2) RepoFactory.create_and_init_dir
same as above but create the directory intended for the repository
before initializing it. For this to work the LocalConnector and
the SSHConnector classes had to be adapted to allow directory
creation of the parent directory (i.e. mkdir('.')). As a
consequence the LocalConnector no longer checks if a path actually
exists when trying to connect to it.